Input: Made the numbers on the display
I was looping through all of the numbersbuttons. Later added eventListener. And then throught data["value"] made the numbers visible on the display. The += in the code is for being able to click on the button as many times as you want. As an example: If you click on button number 1 two times it will show 11.